requirements.txt
django==1.7.1
python-dateutil
gopro
djangorestframework
